requirements.txt
dash>=2.14.0
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.0
gunicorn>=21.2.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.30.0
python-dotenv>=1.0.0

